fix nip42 handling
Currently, filters for kind 4 events are not allowed
for unathenticated clients.  However, when a filter does
not specify kind 4, (e.g. just asks for recent events)
kind 4 events are still returned.

This defeats the point of nip42 since it reveals message
metadata to anyone willing to construct the right filters.

This fix is simple, just copies the same nip42 logic
used to forbid certain filters and applies it to every event
from every filter.